requirement.txt
python-dotenv>=0.20.0
boto3>=1.19.3
